#96 moved all test config classes to a separate package in order to exclude them from the component scan. Although this was no real problem now, it might become one in the future, as it happened in planb-revocation service. An alternative solution is of course, to avoid the component scan at all, and manually wire up the production Spring context.
